backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
chromadb==0.4.24
sentence-transformers==2.2.2
pydantic==2.5.0
python-dotenv==1.0.0

# OCR and PDF Processing
easyocr==1.7.0
pdf2image==1.16.3
pillow==10.1.0
pytesseract==0.3.10

# AI Services
google-generativeai==0.3.0
tenacity==8.2.3

# Async Task Queue (Optional but recommended)
celery==5.3.4
redis==5.0.1
